requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.*
PyMuPDF==1.25.*
python-docx==1.1.*
chardet==5.*
transformers==4.47.*
torch --index-url https://download.pytorch.org/whl/cpu
aiofiles==24.*
python-dotenv==1.*
pydantic-settings==2.7.*
sentence-transformers==3.*
yake==0.4.*
spacy==3.*
